+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 
 ### New
 
+* Checks\DidCommandFail added
 * Checks\DidCommandSucceed added
 * Values\CommandResult added